Move ADC configuration into HAL since it is not actually generic
